requirements.txt
fastapi
uvicorn[standard]
python-dotenv
groq
pypinyin
aiofiles
requests # For pypinyin, sometimes needed for resource files
jinja2
python-multipart
gTTS
supabase-py
